requirements.txt
numpy
pandas
scipy
typer
faiss-gpu
h5py
albumentations
augly
tqdm
matplotlib
opencv-contrib-python
Pillow
pytorch-lightning
lightning-bolts
torchmetrics
tensorflow
scikit-learn
torch
torchvision
timm
jupyter
